refactor(subway): replace legacy LineDao and LineService
- remove all unused code including StatementExecutor and UpdateResult
